+++ src/test/groovy/com/rapidminer/gradle/JavaPublishingReleaseIntegrationSpec.groovy
@@ -32,10 +32,12 @@ class JavaPublishingReleaseIntegrationSpec extends AbstractJavaPublishingIntegra
         setupProject(VERSION)
 
         when:
-        ExecutionResult result = runTasksSuccessfully('publishJarPublicationToMavenRepository')
+        runTasksSuccessfully('publishJarPublicationToMavenRepository')
 
         then:
-        checkMavenRepo(VERSION, new ArtifactConfig(publishTests: true, publishSources: true, publishJavaDoc: true, repo: 'releases'))
+        def config = new ArtifactConfig(publishTests: true, publishSources: true, publishJavaDoc: true, repo: 'releases')
+        checkMavenRepo(VERSION, config)
+        checkPOMContent(VERSION, config, PublishingExtension.LicenseType.RM_EULA)
     }
 
     def 'Test source release publishing'() {
@@ -50,10 +52,12 @@ class JavaPublishingReleaseIntegrationSpec extends AbstractJavaPublishingIntegra
         """.stripIndent()
 
         when:
-        ExecutionResult result = runTasksSuccessfully('publishJarPublicationToMavenRepository')
+        runTasksSuccessfully('publishJarPublicationToMavenRepository')
 
         then:
-        checkMavenRepo(VERSION, new ArtifactConfig(publishTests: true, publishSources: false, publishJavaDoc: true, repo: 'releases'))
+        def config = new ArtifactConfig(publishTests: true, publishSources: false, publishJavaDoc: true, repo: 'releases')
+        checkMavenRepo(VERSION, config)
+        checkPOMContent(VERSION, config, PublishingExtension.LicenseType.RM_EULA)
     }
 
     def 'Test release publishing without javadoc'() {
@@ -68,10 +72,12 @@ class JavaPublishingReleaseIntegrationSpec extends AbstractJavaPublishingIntegra
         """.stripIndent()
 
         when:
-        ExecutionResult result = runTasksSuccessfully('publishJarPublicationToMavenRepository')
+        runTasksSuccessfully('publishJarPublicationToMavenRepository')
 
         then:
-        checkMavenRepo(VERSION, new ArtifactConfig(publishTests: true, publishSources: true, publishJavaDoc: false, repo: 'releases'))
+        def config = new ArtifactConfig(publishTests: true, publishSources: true, publishJavaDoc: false, repo: 'releases')
+        checkMavenRepo(VERSION, config)
+        checkPOMContent(VERSION, config, PublishingExtension.LicenseType.RM_EULA)
     }
 
     def 'Test release publishing without tests'() {
@@ -86,10 +92,12 @@ class JavaPublishingReleaseIntegrationSpec extends AbstractJavaPublishingIntegra
         """.stripIndent()
 
         when:
-        ExecutionResult result = runTasksSuccessfully('publishJarPublicationToMavenRepository')
+        runTasksSuccessfully('publishJarPublicationToMavenRepository')
 
         then:
-        checkMavenRepo(VERSION, new ArtifactConfig(publishTests: false, publishSources: true, publishJavaDoc: true, repo: 'releases'))
+        def config = new ArtifactConfig(publishTests: false, publishSources: true, publishJavaDoc: true, repo: 'releases')
+        checkMavenRepo(VERSION, config)
+        checkPOMContent(VERSION, config, PublishingExtension.LicenseType.RM_EULA)
     }
 
     def 'Test release publishing with different repo'() {
@@ -104,10 +112,12 @@ class JavaPublishingReleaseIntegrationSpec extends AbstractJavaPublishingIntegra
         """.stripIndent()
 
         when:
-        ExecutionResult result = runTasksSuccessfully('publishJarPublicationToMavenRepository')
+        runTasksSuccessfully('publishJarPublicationToMavenRepository')
 
         then:
-        checkMavenRepo(VERSION, new ArtifactConfig(publishTests: true, publishSources: true, publishJavaDoc: true, repo: 'public-releases'))
+        def config = new ArtifactConfig(publishTests: true, publishSources: true, publishJavaDoc: true, repo: 'public-releases')
+        checkMavenRepo(VERSION, config)
+        checkPOMContent(VERSION, config, PublishingExtension.LicenseType.RM_EULA)
     }
 
     def 'Test snapshot publishing with added artifact'() {
@@ -127,11 +137,51 @@ class JavaPublishingReleaseIntegrationSpec extends AbstractJavaPublishingIntegra
         """.stripIndent()
 
         when:
-        ExecutionResult result = runTasksSuccessfully('publishJarPublicationToMavenRepository')
+        def result = runTasksSuccessfully('publishJarPublicationToMavenRepository')
 
         then:
+        def config = new ArtifactConfig(publishTests: true, publishSources: true, publishJavaDoc: true, repo: 'releases')
         result.standardOutput.contains('No credentials defined for publication extension. Looking for \'nexusUser\' and \'nexusPassword\' project properties.')
-        checkMavenRepo(VERSION, new ArtifactConfig(publishTests: true, publishSources: true, publishJavaDoc: true, repo: 'releases'), 'utils')
+        checkMavenRepo(VERSION, config, 'utils')
+        checkPOMContent(VERSION, config, PublishingExtension.LicenseType.RM_EULA)
+    }
+
+    def 'Test snapshot publishing with different vendor and URL'() {
+        setup:
+        setupProject(VERSION)
+        buildFile  << """
+        publication {
+            vendor = 'Another Company'
+            vendorUrl = 'www.anothercompany.com'
+        }
+        """.stripIndent()
+
+        when:
+        runTasksSuccessfully('publishJarPublicationToMavenRepository')
+
+        then:
+        def config = new ArtifactConfig(publishTests: true, publishSources: true, publishJavaDoc: true, repo: 'releases')
+        checkMavenRepo(VERSION, config)
+        checkPOMContent(VERSION, config, 'Another Company', 'www.anothercompany.com', PublishingExtension.LicenseType.RM_EULA)
+    }
+
+    def 'Test snapshot publishing with empty vendor and URL'() {
+        setup:
+        setupProject(VERSION)
+        buildFile  << """
+        publication {
+            vendor ''
+            vendorUrl ''
+        }
+        """.stripIndent()
+
+        when:
+        runTasksSuccessfully('publishJarPublicationToMavenRepository')
+
+        then:
+        def config = new ArtifactConfig(publishTests: true, publishSources: true, publishJavaDoc: true, repo: 'releases')
+        checkMavenRepo(VERSION, config)
+        checkPOMContent(VERSION, config, '', '', PublishingExtension.LicenseType.RM_EULA)
     }
 
     @Override